refactor: separate add_storage_capacity (#594)
